+++ src/main/java/com/fundynamic/d2tm/game/entities/predicates/EntityIsWithinAreaOfAbsoluteCoordinates.java
@@ -3,7 +3,7 @@ package com.fundynamic.d2tm.game.entities.predicates;
 
 import com.fundynamic.d2tm.game.entities.Entity;
 import com.fundynamic.d2tm.game.entities.Predicate;
-import com.fundynamic.d2tm.game.entities.Rectangle;
+import com.fundynamic.d2tm.math.Rectangle;
 
 /**
  * This predicate tests that the given entity's absolute pixel map coordinates are within a given Rectangle.
