+++ src/main/java/com/fundynamic/d2tm/game/entities/predicates/PredicateBuilder.java
@@ -3,6 +3,7 @@ package com.fundynamic.d2tm.game.entities.predicates;
 
 import com.fundynamic.d2tm.game.entities.*;
 import com.fundynamic.d2tm.math.Coordinate;
+import com.fundynamic.d2tm.math.Rectangle;
 import com.fundynamic.d2tm.math.Vector2D;
 
 import java.util.LinkedList;
