+++ src/test/java/com/fundynamic/d2tm/game/controls/MovableSelectedMouseTest.java
@@ -2,7 +2,7 @@ package com.fundynamic.d2tm.game.controls;
 
 import com.fundynamic.d2tm.game.AbstractD2TMTest;
 import com.fundynamic.d2tm.game.controls.battlefield.MovableSelectedMouse;
-import com.fundynamic.d2tm.game.entities.EntitiesData;
+import com.fundynamic.d2tm.game.entities.entitiesdata.EntitiesData;
 import com.fundynamic.d2tm.game.entities.Player;
 import com.fundynamic.d2tm.game.entities.units.Unit;
 import com.fundynamic.d2tm.game.map.Cell;
