+++ src/test/java/com/fundynamic/d2tm/game/entities/RectangleTest.java
@@ -1,5 +1,6 @@
 package com.fundynamic.d2tm.game.entities;
 
+import com.fundynamic.d2tm.math.Rectangle;
 import com.fundynamic.d2tm.math.Vector2D;
 import org.junit.Assert;
 import org.junit.Test;
